Add main README and refernce/project directories
This meeting will be primarily a project workshop in which people will
be experimenting with HTML/CSS. They'll be following a template to
create a web page based on Jake Loo's current page.

Much like the CS61A style, they will be filing in piece sections of the
web page, rather than designing it from scratch.